FIX: Prevent MethodMetrics.totalCount overflow

+++ MyPerf4J-Core/src/main/java/cn/myperf4j/core/recorder/AccurateRecorder.java
@@ -13,13 +13,17 @@ import java.util.concurrent.atomic.AtomicIntegerArray;
  */
 
 /**
- * MyPerf4J默认使用的是 RoughRecorder，如果需要使用 AccurateRecorder，则在配置文件里加上 RecorderMode=accurate
+ * 默认使用该类作为 MyPerf4J 的 Recorder
  * 该类用于精确存储某一个方法在指定时间片内的响应时间
  * 为了减小内存占用，利用 数组+Map 的方式:
  * 1、将小于等于 mostTimeThreshold 的响应时间记录在数组中；
  * 2、将大于 mostTimeThreshold 的响应时间记录到 Map 中。
  */
 public class AccurateRecorder extends Recorder {
+
+//    private static final int BUSY_THRESHOLD = 10;
+//
+//    private final AtomicInteger[] busyTimingArr;
 
     private final AtomicIntegerArray timingArr;
 
@@ -29,10 +33,19 @@ public class AccurateRecorder extends Recorder {
 
     private AccurateRecorder(int methodTagId, int mostTimeThreshold, int outThresholdCount) {
         super(methodTagId);
+//        this.busyTimingArr = createBusyTimingArr();
         this.timingArr = new AtomicIntegerArray(mostTimeThreshold + 1);
         this.timingMap = new ConcurrentHashMap<>(MapUtils.getFitCapacity(outThresholdCount));
         this.diffCount = new AtomicInteger(0);
     }
+
+//    private AtomicInteger[] createBusyTimingArr() {
+//        AtomicInteger[] result = new AtomicInteger[BUSY_THRESHOLD];
+//        for (int i = 0; i < BUSY_THRESHOLD; i++) {
+//            result[i] = new AtomicInteger(0);
+//        }
+//        return result;
+//    }
 
     @Override
     public void recordTime(long startNanoTime, long endNanoTime) {
@@ -41,6 +54,14 @@ public class AccurateRecorder extends Recorder {
         }
 
         int elapsedTime = (int) ((endNanoTime - startNanoTime) / 1000000);
+//        if (elapsedTime < BUSY_THRESHOLD) {
+//            int oldValue = busyTimingArr[elapsedTime].incrementAndGet();
+//            if (oldValue <= 0) {
+//                diffCount.incrementAndGet();
+//            }
+//            return;
+//        }
+
         if (elapsedTime < timingArr.length()) {
             int oldValue = timingArr.getAndIncrement(elapsedTime);
             if (oldValue <= 0) {
@@ -64,8 +85,17 @@ public class AccurateRecorder extends Recorder {
     }
 
     @Override
-    public int fillSortedRecords(IntBuf intBuf) {
-        int totalCount = 0;
+    public long fillSortedRecords(IntBuf intBuf) {
+        long totalCount = 0L;
+//        AtomicInteger[] busyTimingArr = this.busyTimingArr;
+//        for (int i = 0; i < busyTimingArr.length; i++) {
+//            int count = busyTimingArr[i].get();
+//            if (count > 0) {
+//                intBuf.write(i, count);
+//                totalCount += count;
+//            }
+//        }
+
         AtomicIntegerArray timingArr = this.timingArr;
         for (int i = 0; i < timingArr.length(); ++i) {
             int count = timingArr.get(i);
@@ -113,6 +143,11 @@ public class AccurateRecorder extends Recorder {
 
     @Override
     public synchronized void resetRecord() {
+//        AtomicInteger[] busyTimingArr = this.busyTimingArr;
+//        for (int i = 0; i < busyTimingArr.length; i++) {
+//            busyTimingArr[i].set(0);
+//        }
+
         AtomicIntegerArray timingArr = this.timingArr;
         for (int i = 0; i < timingArr.length(); ++i) {
             timingArr.set(i, 0);
